requirements.txt
aiohttp==3.10.10
pyjwt==2.9.0